Wait for new 2fa code to be generated
We recently updated the 2fa code in the admin portal and now it
no longer accepts a 2fa code twice, which is the correct behaviour.

This means that the smoke tests have to wait for 30 seconds for a
new 2fa code to be generated when logging out and back in.